back/StackOverflow/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2  
psycopg2-binary==2.9.6
django-cors-headers==4.0.0